Roblox redesign
Finally finished making this disaster, I'll take a break for 2 days before deciding on another project to work on again.

Note: The mobile support on this is ass, 1. if the height is below 800, it will block the logo, 2. if the device is oriented on landscape mode, the website will cease to break.

Good part though is that the pc works the best.